@@ -30,6 +30,14 @@ plugins {
 group = prop("pluginGroup")
 version = prop("pluginVersion")
 
+val platformVersion = prop("platformVersion")
+val ideaVersionInt = when {
+    // e.g. '20XY.Z'
+    platformVersion.length == 6 -> platformVersion.replace(".", "").substring(2).toInt()
+    // e.g. '2XY.ABCDE.12'
+    else -> platformVersion.substringBefore(".").toInt()
+}
+
 val quarkusVersion = prop("quarkusVersion")
 val lsp4mpVersion = prop("lsp4mpVersion")
 val quarkusLsVersion = prop("quarkusLsVersion")
@@ -75,10 +83,16 @@ sourceSets {
 
 dependencies {
     intellijPlatform {
-        create(prop("platformType"), prop("platformVersion"))
+        create(prop("platformType"), platformVersion)
         // Bundled Plugin Dependencies. Uses `platformBundledPlugins` property from the gradle.properties file for bundled IntelliJ Platform plugins.
         val platformBundledPlugins =  ArrayList<String>()
         platformBundledPlugins.addAll(providers.gradleProperty("platformBundledPlugins").map { it.split(',').map(String::trim).filter(String::isNotEmpty) }.get())
+        /*
+         * platformVersion check for JSON breaking changes since 2024.3
+         */
+        if (ideaVersionInt >= 243) {
+            platformBundledPlugins.add("com.intellij.modules.json")
+        }
         bundledPlugins(platformBundledPlugins)
         // Plugin Dependencies. Uses `platformPlugins` property from the gradle.properties file for plugin from JetBrains Marketplace.
         val platformPlugins =  ArrayList<String>()
@@ -151,10 +165,10 @@ dependencies {
     testImplementation(libs.junit)
 }
 
-// Set the JVM language level used to build the project. Use Java 11 for 2020.3+, and Java 17 for 2022.2+.
+// Set the JVM language level used to build the project. Use Java 11 for 2020.3+, and Java 17 for 2022.2+., and Java 21 for 2025.3+.
 kotlin {
     jvmToolchain {
-        languageVersion = JavaLanguageVersion.of(17)
+        languageVersion = JavaLanguageVersion.of(21)
         vendor = JvmVendorSpec.JETBRAINS
     }
 }
